+++ packages/angular-workspace/example-client-app/src/app/app.module.ts
@@ -10,7 +10,9 @@ import { NimbleTextAreaModule, NimbleTextFieldModule, NimbleNumberFieldModule, N
     NimbleCardButtonModule, NimbleDialogModule, NimbleRadioGroupModule, NimbleRadioModule, NimbleSpinnerModule,
     NimbleAnchorModule, NimbleAnchorButtonModule, NimbleAnchorTabModule, NimbleAnchorTabsModule,
     NimbleIconCheckModule, NimbleBannerModule, NimbleAnchorMenuItemModule, NimbleAnchorTreeItemModule, NimbleIconXmarkCheckModule,
-    NimbleListOptionGroupModule } from '@ni/nimble-angular';
+    NimbleListOptionGroupModule,
+    NimbleIconPencilToRectangleModule,
+    NimbleIconMessagesSparkleModule } from '@ni/nimble-angular';
 import { NimbleCardModule } from '@ni/nimble-angular/card';
 import { NimbleChipModule } from '@ni/nimble-angular/chip';
 import { NimbleLabelProviderCoreModule } from '@ni/nimble-angular/label-provider/core';
@@ -35,7 +37,9 @@ import { NimbleRichTextMentionUsersModule } from '@ni/nimble-angular/rich-text-m
 import { OkButtonModule } from 'ok-angular/button/ok-button.module';
 import { SprightChatConversationModule } from '@ni/spright-angular/chat/conversation';
 import { SprightChatInputModule } from '@ni/spright-angular/chat/input';
-import { SprightChatMessageModule } from '@ni/spright-angular/chat/message';
+import { SprightChatMessageInboundModule } from '@ni/spright-angular/chat/message/inbound';
+import { SprightChatMessageOutboundModule } from '@ni/spright-angular/chat/message/outbound';
+import { SprightChatMessageSystemModule } from '@ni/spright-angular/chat/message/system';
 import { SprightRectangleModule } from '@ni/spright-angular/rectangle';
 import { AppComponent } from './app.component';
 import { CustomAppComponent } from './customapp/customapp.component';
@@ -113,10 +117,14 @@ import { CustomAppComponent } from './customapp/customapp.component';
         NimbleRichTextMentionUsersModule,
         NimbleMappingSpinnerModule,
         NimbleMappingEmptyModule,
+        NimbleIconPencilToRectangleModule,
+        NimbleIconMessagesSparkleModule,
         OkButtonModule,
         SprightChatConversationModule,
         SprightChatInputModule,
-        SprightChatMessageModule,
+        SprightChatMessageInboundModule,
+        SprightChatMessageOutboundModule,
+        SprightChatMessageSystemModule,
         SprightRectangleModule,
         RouterModule.forRoot(
             [