+++ packages/storybook/src/spright/chat/conversation/chat-conversation.stories.ts
@@ -5,6 +5,9 @@ import { menuButtonTag } from '@ni/nimble-components/dist/esm/menu-button';
 import { menuTag } from '@ni/nimble-components/dist/esm/menu';
 import { menuItemTag } from '@ni/nimble-components/dist/esm/menu-item';
 import { toggleButtonTag } from '@ni/nimble-components/dist/esm/toggle-button';
+import { toolbarTag } from '@ni/nimble-components/dist/esm/toolbar';
+import { iconPencilToRectangleTag } from '@ni/nimble-components/dist/esm/icons/pencil-to-rectangle';
+import { iconMessagesSparkleTag } from '@ni/nimble-components/dist/esm/icons/messages-sparkle';
 import {
     ChatConversation,
     chatConversationTag
@@ -14,8 +17,9 @@ import {
     chatInputTag
 } from '@ni/spright-components/dist/esm/chat/input';
 import type { ChatInputSendEventDetail } from '@ni/spright-components/dist/esm/chat/input/types';
-import { ChatMessageType } from '@ni/spright-components/dist/esm/chat/message/types';
-import { chatMessageTag } from '@ni/spright-components/dist/esm/chat/message';
+import { chatMessageInboundTag } from '@ni/spright-components/dist/esm/chat/message/inbound';
+import { chatMessageOutboundTag } from '@ni/spright-components/dist/esm/chat/message/outbound';
+import { chatMessageSystemTag } from '@ni/spright-components/dist/esm/chat/message/system';
 import { richTextViewerTag } from '@ni/nimble-components/dist/esm/rich-text/viewer';
 import { spinnerTag } from '@ni/nimble-components/dist/esm/spinner';
 import { iconCopyTextTag } from '@ni/nimble-components/dist/esm/icons/copy-text';
@@ -36,6 +40,7 @@ import { isChromatic } from '../../../utilities/isChromatic';
 interface ChatConversationArgs {
     appearance: keyof typeof ChatConversationAppearance;
     content: string;
+    toolbar: boolean;
     input: boolean;
     conversationRef: ChatConversation;
     sendMessage: (
@@ -59,22 +64,31 @@ export const chatConversation: StoryObj<ChatConversationArgs> = {
             }
         </style>
         <${chatConversationTag} ${ref('conversationRef')} appearance="${x => x.appearance}">
-            <${chatMessageTag} message-type="${() => ChatMessageType.system}">
+            ${when(x => x.toolbar, html<ChatConversationArgs>`
+                <${toolbarTag} slot='toolbar'>
+                    <${iconMessagesSparkleTag} slot="start"></${iconMessagesSparkleTag}>
+                    <${buttonTag} appearance="ghost" slot="end" title="Create new chat" content-hidden>
+                        Create new chat
+                        <${iconPencilToRectangleTag} slot="start"></${iconPencilToRectangleTag}>
+                    </${buttonTag}>
+                </${toolbarTag}>
+            `)}
+            <${chatMessageSystemTag}>
                 To start, press any key.
-            </${chatMessageTag}>
-            <${chatMessageTag} message-type="${() => ChatMessageType.outbound}">
+            </${chatMessageSystemTag}>
+            <${chatMessageOutboundTag}>
                 Where is the Any key?
-            </${chatMessageTag}>
-            <${chatMessageTag} message-type="${() => ChatMessageType.outbound}">
+            </${chatMessageOutboundTag}>
+            <${chatMessageOutboundTag}>
                 <${richTextViewerTag} markdown="${() => markdownExample}"></${richTextViewerTag}>
-            </${chatMessageTag}>
-            <${chatMessageTag} message-type="${() => ChatMessageType.system}">
+            </${chatMessageOutboundTag}>
+            <${chatMessageSystemTag}>
                 <${spinnerTag}
                     style="${isChromatic() ? '--ni-private-spinner-animation-play-state:paused' : ''}"
                     appearance="${() => SpinnerAppearance.accent}"
                 ></${spinnerTag}>
-            </${chatMessageTag}>
-            <${chatMessageTag} message-type="${() => ChatMessageType.inbound}">
+            </${chatMessageSystemTag}>
+            <${chatMessageInboundTag}>
                 <${buttonTag} slot='footer-actions' appearance='ghost' title='Copy' content-hidden>
                     <${iconCopyTextTag} slot='start'></${iconCopyTextTag}>
                     Copy
@@ -107,7 +121,7 @@ export const chatConversation: StoryObj<ChatConversationArgs> = {
                 <${buttonTag} slot='end' appearance='block'>
                     Check core temperature
                 </${buttonTag}>
-            </${chatMessageTag}>
+            </${chatMessageInboundTag}>
             ${when(x => x.input, html<ChatConversationArgs, ChatInput>`
                 <${chatInputTag} slot='input' placeholder='Type a message' send-button-label='Send'
                     @send="${(x2, c2) => x2.sendMessage(c2.event as CustomEvent<ChatInputSendEventDetail>, x2.conversationRef)}"
@@ -128,6 +142,11 @@ export const chatConversation: StoryObj<ChatConversationArgs> = {
                 'The messages to display in the chat conversation. The DOM order of the messages controls their screen order within the conversation (earlier DOM order implies older message).',
             table: { category: apiCategory.slots }
         },
+        toolbar: {
+            description:
+                `A slot to optionally include a \`${toolbarTag}\` which will be displayed on top of the conversation.`,
+            table: { category: apiCategory.slots }
+        },
         input: {
             description: `A slot to optionally include a \`${chatInputTag}\` which will be displayed below the messages.`,
             table: { category: apiCategory.slots }
@@ -139,9 +158,9 @@ export const chatConversation: StoryObj<ChatConversationArgs> = {
     args: {
         appearance: 'default',
         input: true,
+        toolbar: true,
         sendMessage: (event, conversationRef) => {
-            const message = document.createElement(chatMessageTag);
-            message.messageType = ChatMessageType.outbound;
+            const message = document.createElement(chatMessageOutboundTag);
             const span = document.createElement('span');
             span.textContent = event.detail.text;
             // Preserves new lines and trailing spaces that the user entered