+++ packages/storybook/src/spright/chat/input/chat-input-matrix.stories.ts
@@ -30,6 +30,19 @@ const placeholderStates = [
 ] as const;
 type PlaceholderState = (typeof placeholderStates)[number];
 
+const processingStates = [
+    ['', false],
+    ['processing', true]
+] as const;
+type ProcessingState = (typeof processingStates)[number];
+
+const errorStates = [
+    ['empty', ''],
+    ['one line', 'Error description'],
+    ['multi line', `${loremIpsum} ${loremIpsum} ${loremIpsum}`]
+] as const;
+type ErrorState = (typeof errorStates)[number];
+
 const metadata: Meta = {
     title: 'Tests Spright/Chat Input',
     parameters: {
@@ -41,7 +54,9 @@ export default metadata;
 
 const component = (
     [valueLabel, value]: ValueState,
-    [placeholderLabel, placeholder]: PlaceholderState
+    [placeholderLabel, placeholder]: PlaceholderState,
+    [processingLabel, processing]: ProcessingState,
+    [errorLabel, error]: ErrorState
 ): ViewTemplate => html`
     <p 
         style="
@@ -50,17 +65,20 @@ const component = (
         margin-bottom: 0px;
         "
     >    
-        ${valueLabel} value, ${placeholderLabel} placeholder
+        ${valueLabel} value, ${placeholderLabel} placeholder, ${processingLabel}, ${errorLabel} error
     </p>
     <${chatInputTag}
         placeholder="${placeholder}"
         value="${value}"
+        processing="${processing}"
+        error-visible="${error !== ''}"
+        error-text="${error}"
     >
     </${chatInputTag}>
 `;
 
 export const themeMatrix: StoryFn = createMatrixThemeStory(
-    createMatrix(component, [valueStates, placeholderStates])
+    createMatrix(component, [valueStates, placeholderStates, processingStates, errorStates])
 );
 
 export const hidden: StoryFn = createStory(