+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,64 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Tue, 27 Jan 2026 23:05:34 GMT and should not be manually modified. -->
+<!-- This log was last generated on Mon, 16 Feb 2026 18:09:13 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 35.3.0
+
+Mon, 16 Feb 2026 18:09:13 GMT
+
+### Minor changes
+
+- Add more calendar theme-aware tokens for providing multiple variance of event cards ([ni/nimble@695c1db](https://github.com/ni/nimble/commit/695c1db25df1ec50ff3ca64c55172d7106f21067))
+
+## 35.2.3
+
+Wed, 11 Feb 2026 03:47:57 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v8.15.0
+
+## 35.2.2
+
+Thu, 05 Feb 2026 07:19:38 GMT
+
+### Patches
+
+- visual design updates for switch control ([ni/nimble@fb39fd3](https://github.com/ni/nimble/commit/fb39fd3a30350346e342dca1b9d212ee9b581263))
+
+## 35.2.1
+
+Wed, 04 Feb 2026 23:25:16 GMT
+
+### Patches
+
+- Enable two lint rules and fix violations ([ni/nimble@a410090](https://github.com/ni/nimble/commit/a410090ddbc5c82fbabec6dc67850997d65bd756))
+
+## 35.2.0
+
+Wed, 04 Feb 2026 08:13:24 GMT
+
+### Minor changes
+
+- Add Stepper components ([ni/nimble@1143c37](https://github.com/ni/nimble/commit/1143c3764bee89f48c9b5c6af837a882fa6f06a1))
+
+## 35.1.2
+
+Wed, 04 Feb 2026 01:29:10 GMT
+
+### Patches
+
+- Update fast-element and fast-foundation dependencies ([ni/nimble@6d4d58f](https://github.com/ni/nimble/commit/6d4d58fd3901c8c8e956b1b66c147f98e35973fd))
+
+## 35.1.1
+
+Tue, 03 Feb 2026 14:40:29 GMT
+
+### Patches
+
+- Fix import syntax to obey new lint rule ([ni/nimble@aad1c14](https://github.com/ni/nimble/commit/aad1c1447cb1752602f4cae4e455a7abe0556b06))
 
 ## 35.1.0